+++ src/plugins/replication.js
@@ -19,7 +19,7 @@ import Core from '../core';
 import RxCollection from '../rx-collection';
 import RxError from '../rx-error';
 import PouchDB from '../pouch-db';
-import RxDBWatchForChangesPlugin from '../plugins/watch-for-changes';
+import RxDBWatchForChangesPlugin from './watch-for-changes';
 
 // add pouchdb-replication-plugin
 Core.plugin(PouchReplicationPlugin);