+++ src/plugins/server.js
@@ -8,6 +8,14 @@ import Core from '../core';
 import ReplicationPlugin from './replication';
 Core.plugin(ReplicationPlugin);
 
+import RxDBWatchForChangesPlugin from './watch-for-changes';
+Core.plugin(RxDBWatchForChangesPlugin);
+
+// we have to clean up after tests so there is no stupid logging
+// @link https://github.com/pouchdb/pouchdb-server/issues/226
+const PouchdbAllDbs = require('pouchdb-all-dbs');
+PouchdbAllDbs(PouchDB);
+
 const APP_OF_DB = new WeakMap();
 const SERVERS_OF_DB = new WeakMap();
 const DBS_WITH_SERVER = new WeakSet();
@@ -59,7 +67,6 @@ export function spawnServer({
 
 
     // tunnel requests so collection-names can be used as paths
-    // TODO do this for all collections that get created afterwards
     Object.keys(db.collections).forEach(colName => tunnelCollectionPath(db, path, app, colName));
 
     // show error if collection is created afterwards